+++ include/WPA/TypeBasedHeapCloning.h
